Fix test_disabling_room_list
I think this now fails because I've changed the user directory code to
require that users are not deactivated. I think this test was relying on
some magic to sort of create a user based on the test case attribute
"user_id". I don't think it was doing it properly by actually
registering a user proper.