Move 'remove deleted devices' migration to latest schema version
This migration file was included as part of an old PR, and at the time of the PR's creation,
the latest schema version was 64. When the PR was merged, the latest version was 65, however
the migration was not moved. This meant that Synapse's on schema version 65 would not run
the migration after updating.

Move the migration file to the latest schema, 65. It is OK if this SQL runs twice, as it
simply kicks off a background table cleanup process.