Pull in StatsStore
I need StateDeltasStore for _curr_state_delta_stream_cache.
I also need the `clock` attribute, which pathologically only seems to be
defined on StatsStore. StatsStore is a subclass of StateDeltasStore, so
let's just use that to keep the list of parents minimal.

To be explicit: I think it's very very odd that only StatsStore provides
a `clock`. I think pulling it in here is silly: why should
CacheInvalidationWorkerStore care about stats? Consider this a protest
commit to raise attention to the bizarre status quo.