Simply calling execute_values in order to appease mypy
mypy was getting upset as it thought 'fetch' was being passed multiple times
to execute_values. This cannot happen, as fetch would never be part of 'args',
and thus seems like a bug in mypy.

This code was a bit strange anyhow. Why collect values with *x only to
expand them again immediately afterwards when calling execute_values?

The code is simpler now IMO, and appeases mypy.