Revert changes to search_user_dir
We no longer use this change, and it actually causes the function to
fail in the case that the query is empty.

I think the reason we needed this in the past was that the API server
queried Synapse directly, but now it has its own database, so these
changes can be reverted.

This brings the synapse/storage/databases/main/user_directory.py in line
with upstream. Best way to test this is to run

    git diff v1.47.0

which will give no diff for this file.