Fix `LruCache` corruption bug with a `size_callback` that can return 0
When all entries in an `LruCache` have a size of 0 according to the
provided `size_callback`, and `drop_from_cache` is called on a cache
node, the node would be unlinked from the LRU linked list but remain in
the cache dictionary. An assertion would be later be tripped due to the
inconsistency.

Use a strict `is None` check instead when unwrapping the weak reference.