Fix query rejections and edge cases
See

 - https://github.com/matrix-org/synapse/pull/9445#discussion_r757090973
 - https://github.com/matrix-org/synapse/pull/9445#discussion_r757092652

This commit does make one of the federation Complement tests fail
but I think it's because our backward extremity query is weird
and doesn't check the event in questions prev_events. Currently
we only check the event itself which won't be a backward extremity
because we already have the event and set the prev_events as the
new backward extremities.