Remove DomainRuleChecker module (#113)
The module has now moved to https://github.com/matrix-org/synapse-domain-rule-checker

__Note to ops__:

To deploy this change, the module needs to be installed in addition to Synapse (from the aforementioned repo, or the PyPI project mentioned its readme). The configuration doesn't change besides a) `default` is renamed into `can_invite_if_not_in_domain_mapping` and b) the module's configuration has moved to the `modules` section of the configuration file (though that was already the case as of https://github.com/matrix-org/synapse-dinsic/pull/108).