Allow tracking puppeted users for MAU
Currently when puppeting another user, the user doing the puppeting is tracked for client ip's and MAU (if configured).

When tracking MAU is important, it becomes necessary to be possible to also track the client IP's and MAU of puppeted users. As an example a client that manages user creation and creation of tokens via the Synapse admin API, passing those tokens for the client to use.

This PR adds optional configuration to enable tracking of puppeted users into monthly active users. The default behaviour stays the same.

Signed-off-by: Jason Robinson <jasonr@matrix.org>