Optionally use an on-disk sqlite db in tests
When debugging a test it is sometimes useful to inspect the state of the
DB. This is not easy when the db is in-memory: one cannot attach the
sqlite CLI to another process's DB.

With this change, if SYNAPSE_TEST_SQLITE_DB_LOCATION is set, we try to
write the DB to disk at that location (deleting any existing file as
necessary).

Quick and dirty.