Beeper hack: disable tracking of AS last txn IDs.
The update to `application_services_state` often causes serialization
errors as the AS handler also updates the stream position. The check
here exists only to display a warning, which we see from time to time,
but has no real affect on the processing or handling of AS transactions
thus making it pretty much redundant.