Make matches_user_in_member_list method private; update refs
References to this method wanted to check if an AS should receive events
from a room. They were only checking to see if the AS had a user in the
room, whereas the AS should also see those events if it has registered
the room's ID or one of its aliases in the AS' relevant namespace.