media/{download,thumbnail}: add support for stall parameter
The stall will default to 20s if MSC2246 is not enabled.

For remote media, the max_stall_ms parameter through to remote server.

This way, if servers over federation attempt to request the media, the
request will only 404 if the media is large (or the client doesn't
upload as soon as it can).

Signed-off-by: Sumner Evans <sumner@beeper.com>