add default_power_level_content_override config option.
Lets you override the default power levels for rooms created on this server, per
room creation preset.

Useful if you know that your users need special permissions in rooms
that they create (e.g. to send particular types of state events without
needing an elevated power level).  This takes the same shape as the
power_level_content_override parameter in the /createRoom API, but
is applied before that parameter.

This is something of a workaround in the absence of MSC3779 or MSC3761.